Replaced np.int and np.uint with np.int32 and np.uint32

+++ model/ns3gym/ns3gym/ns3env.py
@@ -112,9 +112,9 @@ class Ns3ZmqBridge(object):
             mtype = boxSpacePb.dtype
 
             if mtype == pb.INT:
-                mtype = np.int
+                mtype = np.int32
             elif mtype == pb.UINT:
-                mtype = np.uint
+                mtype = np.uint32
             elif mtype == pb.DOUBLE:
                 mtype = np.float64
             else:
